# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Online Stock",
    "description": "Welcome to the Documentation of the Online Stock API!\n\nHere, we provide detailed information to facilitate the integration and optimize the usage of our Online Stock API, whether you are a developer or a business owner. Don't forget to perform token verification to access all functions of the API.",
    "contact": {
      "email": "gui.edu0380@gmail.com"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "https://localhost:3000",
      "description": "local server"
    }
  ],
...
  "paths": {
...
    "/table/{tableName}": {
      "delete": {
        "tags": [
          "stock"
        ],
        "summary": "Delete Table",
        "description": "Delete a table if it is empty",
        "operationId": "deleteTable",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "description": "the name of table",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HttpDeleteTableResponse200"
          },
          "400": {
            "$ref": "#/components/responses/HttpDeleteTableResponse400"
          },
          "404": {
            "$ref": "#/components/responses/HttpDeleteTableResponse404"
          },
          "500": {
            "$ref": "#/components/responses/Http500"
          }
        }
      },
      "post": {
        "tags": [
          "stock"
        ],
        "summary": "Add products to the table",
        "description": "Add one product in the table",
        "operationId": "addProduct",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "description": "the name of table",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Create a new product in the table, price in cents",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/insertProducts"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/HttpInsertProductsResponse201"
          },
          "405": {
            "$ref": "#/components/responses/HttpInsertProductsResponse404"
          },
          "500": {
            "$ref": "#/components/responses/Http500"
          }
        }
      },
      "get": {
        "tags": [
          "stock"
        ],
        "summary": "Listing products",
        "description": "list all the products, or specify a product with a query",
        "operationId": "findPetsByStatus",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "description": "The name of table",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productName",
            "in": "query",
            "description": "The name of specific product",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HttpListingProductsResponse200q"
          },
          "404": {
            "$ref": "#/components/responses/HttpListingProductsResponse404"
          },
          "500": {
            "$ref": "#/components/responses/Http500"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "insertProducts": {
        "type": "object",
        "properties": {
          "product_name": {
            "type": "string",
            "example": "product 1"
          },
          "amount_stock": {
            "type": "integer",
            "example": 100
          },
          "price": {
            "type": "integer",
            "example": 1000
          },
          "description": {
            "type": "string",
            "example": "a description for product 1"
          },
          "link": {
            "type": "string",
            "example": "https://linkforproduct.com/"
          }
        }
      },
      "listingProducts": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "integer",
            "example": 1
          },
          "product_name": {
            "type": "string",
            "example": "product 1"
          },
          "amount_stock": {
            "type": "integer",
            "example": 100
          },
          "price": {
            "type": "integer",
            "example": 1000
          },
          "description": {
            "type": "string",
            "example": "a description for product 1"
          },
          "link": {
            "type": "string",
            "example": "https://linkforproduct.com/"
          }
        }
      },
      "listingProductsWithTable": {
        "type": "object",
        "properties": {
          "table": {
            "type": "string",
            "example": "Table Name"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/listingProducts"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "HttpInsertProductsResponse201": {
        "description": "Created",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Product has been created."
                }
              }
            }
          }
        }
      },
      "HttpInsertProductsResponse404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Table \"New Table\" not exists."
                }
              }
            }
          }
        }
      },
      "HttpListingProductsResponse200q": {
        "description": "Sucess",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/listingProductsWithTable"
                },
                {
                  "$ref": "#/components/schemas/listingProducts"
                }
              ]
            }
          }
        }
      },
...
      "HttpListingProductsResponse404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Table or Product not found"
                }
              }
            }
          }
        }
      },
      "HttpDeleteTableResponse200": {
        "description": "Sucess",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "The table has been deleted."
                }
              }
            }
          }
        }
      },
      "HttpDeleteTableResponse400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "The table must be empty to be deleted"
                }
              }
            }
          }
        }
      },
      "HttpDeleteTableResponse404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Table not found"
                }
              }
            }
          }
        }
      },
...
      "Http500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error."
                }
              }
            }
          }
        }
      }
    }
  }
}